Revisión de tipo

+++ agendaMLG/src/main/java/grupog/agendamlg/Usuario.java
@@ -46,8 +46,8 @@ public class Usuario implements Serializable {
     private String password_hash;
     @Column(name="sal", nullable=false)
     private String sal;
-    private Long longitud;
-    private Long latitud;
+    private Double longitud;
+    private Double latitud;
     
     @OneToMany(cascade=CascadeType.ALL)
     private List<Comentario> comentarios;
@@ -139,19 +139,19 @@ public class Usuario implements Serializable {
         this.sal = sal;
     }
 
-    public Long getLongitud() {
+    public Double getLongitud() {
         return longitud;
     }
 
-    public void setLongitud(Long longitud) {
+    public void setLongitud(Double longitud) {
         this.longitud = longitud;
     }
 
-    public Long getLatitud() {
+    public Double getLatitud() {
         return latitud;
     }
 
-    public void setLatitud(Long latitud) {
+    public void setLatitud(Double latitud) {
         this.latitud = latitud;
     }
